♻️ refactor: 회식별 참석자 수 조회 N + 1 문제 해결
- 각 회식의 참석자 수를 Map 형태로 만들어서 한 번의 쿼리로 수행

+++ src/main/java/com/team8/damo/service/response/DiningResponse.java
@@ -11,9 +11,9 @@ public record DiningResponse(
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm")
     LocalDateTime diningDate,
     DiningStatus status,
-    int diningParticipantsCount
+    Long diningParticipantsCount
 ) {
-    public static DiningResponse of(Dining dining, int participantsCount) {
+    public static DiningResponse of(Dining dining, Long participantsCount) {
         return new DiningResponse(
             dining.getId(),
             dining.getDiningDate(),
